Don't put tokens in req.session. Just don't
This has been the source of expired token issues and causing blank pages
and redirects. Turns out we don't really need to keep the token around
in the session anyway, so let's just not do it.

Access tokens are taken from a cookie, if available, validated, and
stored on the request object so other handlers can access it if needed.